seperate IdxRange and ThreadSpace
